#434 fulltextsearchパッケージに移動
