+++ iplass-admin/src/main/java/org/iplass/adminconsole/server/base/service/AdminEntityManager.java
@@ -35,7 +35,6 @@ import org.iplass.mtp.entity.DeleteCondition;
 import org.iplass.mtp.entity.DeleteOption;
 import org.iplass.mtp.entity.Entity;
 import org.iplass.mtp.entity.EntityManager;
-import org.iplass.mtp.entity.FulltextSearchOption;
 import org.iplass.mtp.entity.InsertOption;
 import org.iplass.mtp.entity.LoadOption;
 import org.iplass.mtp.entity.SearchOption;
@@ -44,6 +43,7 @@ import org.iplass.mtp.entity.UpdateCondition;
 import org.iplass.mtp.entity.UpdateOption;
 import org.iplass.mtp.entity.ValidateResult;
 import org.iplass.mtp.entity.bulkupdate.BulkUpdatable;
+import org.iplass.mtp.entity.fulltextsearch.FulltextSearchOption;
 import org.iplass.mtp.entity.query.Query;
 import org.iplass.mtp.impl.entity.auditlog.AuditLogInterceptor;
 import org.iplass.mtp.spi.ServiceRegistry;
