+++ iplass-core/src/main/java/org/iplass/mtp/entity/fulltextsearch/FulltextSearchOption.java
@@ -18,13 +18,10 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.iplass.mtp.entity;
+package org.iplass.mtp.entity.fulltextsearch;
 
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
-
-import org.iplass.mtp.entity.query.OrderBy;
 
 /**
  * 全文検索時のオプション
@@ -47,38 +44,4 @@ public class FulltextSearchOption {
 		this.conditions = conditions;
 	}
 
-	public static class FulltextSearchCondition {
-
-		private List<String> properties;
-
-		private OrderBy order;
-
-		public FulltextSearchCondition() {
-		}
-
-		public FulltextSearchCondition(List<String> properties) {
-			this(properties, null);
-		}
-
-		public FulltextSearchCondition(List<String> properties, OrderBy order) {
-			this.properties = properties;
-			this.order = order;
-		}
-
-		public List<String> getProperties() {
-			return properties;
-		}
-
-		public void setProperties(List<String> properties) {
-			this.properties = properties;
-		}
-
-		public OrderBy getOrder() {
-			return order;
-		}
-
-		public void setOrder(OrderBy order) {
-			this.order = order;
-		}
-	}
 }
