+++ iplass-core/src/main/java/org/iplass/mtp/impl/entity/EntityManagerImpl.java
@@ -46,7 +46,6 @@ import org.iplass.mtp.entity.EntityConcurrentUpdateException;
 import org.iplass.mtp.entity.EntityManager;
 import org.iplass.mtp.entity.EntityRuntimeException;
 import org.iplass.mtp.entity.EntityValidationException;
-import org.iplass.mtp.entity.FulltextSearchOption;
 import org.iplass.mtp.entity.InsertOption;
 import org.iplass.mtp.entity.LoadOption;
 import org.iplass.mtp.entity.SearchOption;
@@ -67,6 +66,7 @@ import org.iplass.mtp.entity.definition.properties.BinaryProperty;
 import org.iplass.mtp.entity.definition.properties.ReferenceProperty;
 import org.iplass.mtp.entity.definition.properties.ReferenceType;
 import org.iplass.mtp.entity.definition.properties.StringProperty;
+import org.iplass.mtp.entity.fulltextsearch.FulltextSearchOption;
 import org.iplass.mtp.entity.interceptor.InvocationType;
 import org.iplass.mtp.entity.query.Limit;
 import org.iplass.mtp.entity.query.Query;
