+++ iplass-core/src/main/java/org/iplass/mtp/impl/fulltextsearch/FulltextSearchLuceneService.java
@@ -81,11 +81,11 @@ import org.iplass.mtp.definition.DefinitionSummary;
 import org.iplass.mtp.entity.BinaryReference;
 import org.iplass.mtp.entity.Entity;
 import org.iplass.mtp.entity.EntityManager;
-import org.iplass.mtp.entity.FulltextSearchOption;
-import org.iplass.mtp.entity.FulltextSearchOption.FulltextSearchCondition;
 import org.iplass.mtp.entity.SearchResult;
 import org.iplass.mtp.entity.definition.EntityDefinition;
 import org.iplass.mtp.entity.definition.EntityDefinitionManager;
+import org.iplass.mtp.entity.fulltextsearch.FulltextSearchCondition;
+import org.iplass.mtp.entity.fulltextsearch.FulltextSearchOption;
 import org.iplass.mtp.entity.fulltextsearch.FulltextSearchRuntimeException;
 import org.iplass.mtp.entity.query.OrderBy;
 import org.iplass.mtp.entity.query.Query;
