+++ iplass-gem/src/main/java/org/iplass/gem/command/fulltext/FullTextSearchCommand.java
@@ -27,7 +27,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -52,13 +51,13 @@ import org.iplass.mtp.command.annotation.CommandClass;
 import org.iplass.mtp.command.annotation.webapi.WebApi;
 import org.iplass.mtp.entity.Entity;
 import org.iplass.mtp.entity.EntityManager;
-import org.iplass.mtp.entity.FulltextSearchOption;
-import org.iplass.mtp.entity.FulltextSearchOption.FulltextSearchCondition;
 import org.iplass.mtp.entity.definition.EntityDefinition;
 import org.iplass.mtp.entity.definition.EntityDefinitionManager;
 import org.iplass.mtp.entity.definition.PropertyDefinition;
 import org.iplass.mtp.entity.definition.properties.ReferenceProperty;
+import org.iplass.mtp.entity.fulltextsearch.FulltextSearchCondition;
 import org.iplass.mtp.entity.fulltextsearch.FulltextSearchManager;
+import org.iplass.mtp.entity.fulltextsearch.FulltextSearchOption;
 import org.iplass.mtp.entity.fulltextsearch.FulltextSearchRuntimeException;
 import org.iplass.mtp.entity.permission.EntityPermission;
 import org.iplass.mtp.entity.query.OrderBy;
